pyproject.toml
[project]
name = "receipt_processor"
version = "0.1.0"
description = ""
authors = [
    {name = "Amandeep Singh",email = "amansingh78622@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "spacy (>=3.8.6,<4.0.0)",
    "flask-sqlalchemy (==3.1.1)",
    "werkzeug (==2.3.7)",
    "pypdf2 (==3.0.1)",
    "pytesseract (==0.3.10)",
    "pdf2image (==1.16.3)",
    "python-dateutil (==2.8.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
